fixed: warning
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/trunk/XBMC@6382 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ guilib/GUIRSSControl.cpp
@@ -112,7 +112,7 @@ void CGUIRSSControl::RenderText()
 
   if (m_scrollInfo.initialWait == -1)
   {
-    m_scrollInfo.initialWait = (g_graphicsContext.GetVideoResolution()<6?60.f:50.f)/g_graphicsContext.GetWidth()*720;
+    m_scrollInfo.initialWait = (int)(g_graphicsContext.GetVideoResolution()<6?60.f:50.f)/g_graphicsContext.GetWidth()*720;
     m_scrollInfo.Reset();
   }
